backend/requirements.txt
FastAPI
uvicorn
sqlalchemy
pydantic
databases
asyncpg
httpx
python-dotenv
python-jose[cryptography]
passlib[bcrypt]
python-multipart
pydantic[email]